refactor to avoid foot cannon
avoid defining the valid sysinfo block details as an attribute on the
config class to prevent confusion about what should appear with the
template helper.

refactor the validation to be more strict on bad block names or entries
with specific supported names, retaining the warning on empty entries
being skipped.