driver: fix qemu_use_session
Detection of network connections via system is broken, but directly querying libvirt is working properly. Using the system connection with qemu_use_session is enabled will lead to polkit trying to acquire root privileges, which is unavailable in some setups. The client is nowadays capable of enumerating the network devices as unprivileged user.
Fixes #1808.